Struct twox_hash::XxHash32

pub struct XxHash32 { /* fields omitted */ }

Calculates the 32-bit hash. Care should be taken when using this hash.

Although this struct implements Hasher, it only calculates a 32-bit number, leaving the upper bits as 0. This means it is unlikely to be correct to use this in places like a HashMap.

Methods

impl XxHash

pub fn with_seed(seed: u32) -> XxHash

Constructs the hash with an initial seed