Struct twox_hash::XxHash32

pub struct XxHash32 { /* private fields */ }

Calculates the 32-bit hash. Care should be taken when using this hash.

Although this struct implements Hasher, it only calculates a 32-bit number, leaving the upper bits as 0. This means it is unlikely to be correct to use this in places like a HashMap.

Implementations

impl XxHash32

pub fn with_seed(seed: u32) -> XxHash32

Constructs the hash with an initial seed

pub fn seed(&self) -> u32

pub fn total_len(&self) -> u32

Get the total number of bytes hashed, truncated to 32 bits. For the full 64-bit byte count, use total_len_64

pub fn total_len_64(&self) -> u64

Get the total number of bytes hashed.