twox_hash

Module xxh3

The in-progress XXH3 algorithm.

Please read the notes in original implementation to learn about when to use these algorithms. Specifically, the version of code this crate reproduces says:

The algorithm is currently in development, meaning its return values might still change in future versions. However, the API is stable, and can be used in production, typically for generation of ephemeral hashes (produced and consumed in same session).

Structs

Hash64

Calculates the 64-bit hash.

Hash128

Calculates the 128-bit hash.

Traits

HasherExt

Functions

hash64

hash64_with_secret

hash64_with_seed

hash128

hash128_with_secret

hash128_with_seed